# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: HR Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /employees:
    get:
      operationId: listEmployees
      tags: [employees]
      summary: List all employees
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: department
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: designation
          schema:
            type: string
      responses:
        "200":
          description: List of employees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      operationId: createEmployee
      tags: [employees]
      summary: Create new employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmployeeBody"
      responses:
        "201":
          description: Created employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/{id}:
    get:
      operationId: getEmployee
      tags: [employees]
      summary: Get employee by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "404":
          description: Not found
    patch:
      operationId: updateEmployee
      tags: [employees]
      summary: Update employee
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEmployeeBody"
      responses:
        "200":
          description: Updated employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    delete:
      operationId: deleteEmployee
      tags: [employees]
      summary: Delete employee
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /employees/{id}/documents:
    get:
      operationId: listEmployeeDocuments
      tags: [employees]
      summary: List employee documents
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmployeeDocument"
    post:
      operationId: uploadEmployeeDocument
      tags: [employees]
      summary: Upload employee document
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadDocumentBody"
      responses:
        "201":
          description: Created document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeDocument"
  /employees/{id}/history:
    get:
      operationId: getEmployeeHistory
      tags: [employees]
      summary: Get promotion/transfer history
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: History
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmployeeHistory"
    post:
      operationId: addEmployeeHistory
      tags: [employees]
      summary: Add history entry
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddEmployeeHistoryBody"
      responses:
        "201":
          description: History entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeHistory"
...
  /automations/rules/{id}/toggle:
    post:
      operationId: toggleAutomationRule
      tags: [automations]
      summary: Toggle automation rule on/off
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToggleAutomationBody"
      responses:
        "200":
          description: Updated rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutomationRule"
  /automations/logs:
    get:
      operationId: listAutomationLogs
      tags: [automations]
      summary: List automation logs
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AutomationLog"
  /automations/email-templates:
    get:
      operationId: listEmailTemplates
      tags: [automations]
      summary: List email templates
      responses:
        "200":
          description: Templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmailTemplate"
    post:
      operationId: createEmailTemplate
      tags: [automations]
      summary: Create email template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmailTemplateBody"
      responses:
        "201":
          description: Template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailTemplate"
  /automations/email-templates/{id}:
    patch:
      operationId: updateEmailTemplate
      tags: [automations]
      summary: Update email template
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmailTemplateBody"
      responses:
        "200":
          description: Template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailTemplate"
components:
  schemas:
...
    Employee:
      type: object
      properties:
        id:
          type: string
        employeeCode:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        dateOfBirth:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        emergencyContact:
          type: ["string", "null"]
        emergencyPhone:
          type: ["string", "null"]
        departmentId:
          type: ["string", "null"]
        departmentName:
          type: ["string", "null"]
        designationId:
          type: ["string", "null"]
        designationName:
          type: ["string", "null"]
        reportingManagerId:
          type: ["string", "null"]
        reportingManagerName:
          type: ["string", "null"]
        employmentType:
          type: string
        joiningDate:
          type: string
        probationEndDate:
          type: ["string", "null"]
        status:
          type: string
        lastWorkingDay:
          type: ["string", "null"]
        resignationDate:
          type: ["string", "null"]
        fnfStatus:
          type: ["string", "null"]
        userId:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, employeeCode, firstName, lastName, email, employmentType, joiningDate, status, createdAt, updatedAt]
    CreateEmployeeBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        dateOfBirth:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        emergencyContact:
          type: ["string", "null"]
        emergencyPhone:
          type: ["string", "null"]
        departmentId:
          type: ["string", "null"]
        designationId:
          type: ["string", "null"]
        reportingManagerId:
          type: ["string", "null"]
        employmentType:
          type: string
        joiningDate:
          type: string
        probationEndDate:
          type: ["string", "null"]
        status:
          type: string
      required: [firstName, lastName, email, employmentType, joiningDate, status]
    UpdateEmployeeBody:
      type: object
      properties:
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        dateOfBirth:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        emergencyContact:
          type: ["string", "null"]
        emergencyPhone:
          type: ["string", "null"]
        departmentId:
          type: ["string", "null"]
        designationId:
          type: ["string", "null"]
        reportingManagerId:
          type: ["string", "null"]
        employmentType:
          type: ["string", "null"]
        joiningDate:
          type: ["string", "null"]
        probationEndDate:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        lastWorkingDay:
          type: ["string", "null"]
        resignationDate:
          type: ["string", "null"]
        fnfStatus:
          type: ["string", "null"]
    EmployeeDocument:
      type: object
      properties:
        id:
          type: string
        employeeId:
          type: string
        documentType:
          type: string
        fileName:
          type: string
        fileUrl:
          type: string
        fileSize:
          type: ["integer", "null"]
        uploadedAt:
          type: string
        verifiedAt:
          type: ["string", "null"]
        verifiedBy:
          type: ["string", "null"]
      required: [id, employeeId, documentType, fileName, fileUrl, uploadedAt]
    UploadDocumentBody:
      type: object
      properties:
        documentType:
          type: string
        fileName:
          type: string
        fileUrl:
          type: string
        fileSize:
          type: ["integer", "null"]
      required: [documentType, fileName, fileUrl]
    EmployeeHistory:
      type: object
      properties:
        id:
          type: string
        employeeId:
          type: string
        changeType:
          type: string
        previousValue:
          type: ["string", "null"]
        newValue:
          type: ["string", "null"]
        effectiveDate:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        createdBy:
          type: ["string", "null"]
      required: [id, employeeId, changeType, effectiveDate, createdAt]
    AddEmployeeHistoryBody:
      type: object
      properties:
        changeType:
          type: string
        previousValue:
          type: ["string", "null"]
        newValue:
          type: ["string", "null"]
        effectiveDate:
          type: string
        notes:
          type: ["string", "null"]
      required: [changeType, effectiveDate]
...
    AutomationRule:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        code:
          type: string
        triggerType:
          type: string
        triggerEvent:
          type: ["string", "null"]
        cronExpr:
          type: ["string", "null"]
        templateId:
          type: string
        templateName:
          type: string
        recipients:
          type: string
        isActive:
          type: boolean
        lastRunAt:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, name, code, triggerType, templateId, templateName, recipients, isActive, createdAt]
    ToggleAutomationBody:
      type: object
      properties:
        isActive:
          type: boolean
      required: [isActive]
    AutomationLog:
      type: object
      properties:
        id:
          type: string
        ruleId:
          type: string
        ruleName:
          type: string
        employeeId:
          type: ["string", "null"]
        employeeName:
          type: ["string", "null"]
        templateCode:
          type: string
        recipientEmail:
          type: string
        status:
          type: string
        errorMessage:
          type: ["string", "null"]
        sentAt:
          type: string
      required: [id, ruleId, ruleName, templateCode, recipientEmail, status, sentAt]
    EmailTemplate:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        name:
          type: string
        subject:
          type: string
        bodyHtml:
          type: string
        variables:
          type: array
          items:
            type: string
        isActive:
          type: boolean
        updatedAt:
          type: string
      required: [id, code, name, subject, bodyHtml, variables, isActive, updatedAt]
    CreateEmailTemplateBody:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        subject:
          type: string
        bodyHtml:
          type: string
        variables:
          type: array
          items:
            type: string
        isActive:
          type: boolean
      required: [code, name, subject, bodyHtml]